+++ CHANGES.md
@@ -11,6 +11,7 @@ Release Notes.
 * Add recover to goroutine to prevent unexpected panics.
 * Add mutex to fix some data race. 
 * Replace external `goapi` dependency with in-repo generated protocols. 
+* Support pprof profiling. 
 #### Plugins
 
 #### Documentation